Curs_8/02_multimodal_image/requirements.txt
python-dotenv
flask
langchain-google-genai
langchain-community
bs4